+++ emx/lib/str/memccpy.c
@@ -2,9 +2,9 @@
 
 #include <string.h>
 
-void *memccpy (void *s1, const void *s2, int c, int n)
+void *memccpy (void *s1, const void *s2, int c, size_t n)
     {
-    int i;
+    size_t i;
   
     for (i = 0; i < n; ++i)
         {
