+++ emx/lib/str/memcpy.c
@@ -2,7 +2,7 @@
 
 #include <string.h>
 
-void *memcpy (void *s1, const void *s2, int n)
+void *memcpy (void *s1, const void *s2, size_t n)
     {
     int i;
