+++ emx/lib/str/memmove.c
@@ -2,7 +2,7 @@
 
 #include <string.h>
 
-void *memmove (void *s1, const void *s2, int n)
+void *memmove (void *s1, const void *s2, size_t n)
     {
     int i;
